// front/src/providers/api-client.generated/model/addressDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressDto { 
    id?: string;
    lineOne?: string;
    lineTwo?: string;
    label?: string;
    department?: string;
    postalCode?: string;
    city?: string;
    country?: string;
    candidateId?: string;
    customerId?: string;
    candidateApplicationId?: string;
    jobReferenceId?: string;
}