// front/src/providers/api-client.generated/model/candidateJobDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobReferenceDto } from './jobReferenceDto';
import { CandidateDto } from './candidateDto';
import { AppValueDto } from './appValueDto';


export interface CandidateJobDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    candidateId?: string;
    candidate?: CandidateDto;
    candidateResumeId?: string;
    job?: AppValueDto;
    jobId?: string;
    experienceStartDate?: Date;
    experienceEndDate?: Date;
    showMonthInResume?: boolean;
    postTitle?: string;
    postDescription?: string;
    postResponsability?: string;
    employer?: string;
    inActivity?: boolean;
    leavingReason?: string;
    jobReference?: JobReferenceDto;
    jobReferenceId?: string;
    employerProfileId?: string;
    employerProfile?: AppValueDto;
    jobName?: string;
    jobDescription?: string;
    status?: CandidateJobDto.StatusEnum;
    type?: CandidateJobDto.TypeEnum;
}
export namespace CandidateJobDto {
    export type StatusEnum = 'VALIDATED' | 'REFUSED' | 'PENDING';
    export const StatusEnum = {
        Validated: 'VALIDATED' as StatusEnum,
        Refused: 'REFUSED' as StatusEnum,
        Pending: 'PENDING' as StatusEnum
    };
    export type TypeEnum = 'JOB' | 'EDUCATION';
    export const TypeEnum = {
        Job: 'JOB' as TypeEnum,
        Education: 'EDUCATION' as TypeEnum
    };
}