// front/src/providers/api-client.generated/model/jobReferenceDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressDto } from './addressDto';
import { AppValueDto } from './appValueDto';


export interface JobReferenceDto { 
    id: string;
    email?: string;
    phone?: string;
    addresses?: Array<AddressDto>;
    jobRefFunctionId: string;
    jobRefFunction?: AppValueDto;
    otherFunction?: string;
    candidateAcceptContact: boolean;
    isPrivatePerson: boolean;
    isCompany: boolean;
    privatePersonFirstName?: string;
    privatePersonLastName?: string;
    companyName?: string;
    note?: string;
    disabled?: boolean;
    customerHasBeenCreated: boolean;
    contactFullName?: string;
    jobReferenceId?: string;
}